added afterstorage flashcards for placing and removing from oven

+++ services/dashboard_service.py
@@ -14,8 +14,14 @@ class DashboardService:
         }
 
     def place_afterstorage(self, sample_id, oven_location):
+        if self.repository.afterstorage_place_exists(sample_id):
+            raise ValueError("Deze Batch is al in de oven geplaatst!")
+
         self.repository.place_afterstorage(sample_id, oven_location)
 
-    def remove_afterstorage(self, afterstorage_id):
+    def remove_afterstorage(self, sample_id, afterstorage_id):
+        if self.repository.afterstorage_remove_exists(sample_id):
+            raise ValueError("Deze Batch is al uit de oven gehaald!")
+            
         self.repository.remove_afterstorage(afterstorage_id)
